refactor: separate admin account and user controllers
